food_tracker/requirements.txt
streamlit>=1.55.0
pillow>=10.0.0
pandas>=2.0.0
opencv-python-headless>=4.8.0
pyzbar>=0.1.9
requests>=2.31.0
